Separating out GameObject and RFrame objects
To help with access speed to Rframe objects, the arrays that hold the vertices and struts and their Unity GameObjects are held in their own arrays.
This means that when a strut or vertex is needed, Unity isn't haven't to do a GetComponent call.
Already this seems to have sped things up a bit.
This is leading to implementing "headless" visual code and makes portability to other code languges easier.