Double Vertex Implemetation
Added a duplicate vertex for the Behaviour Mods with a strut connecting to the original vertex.
This allows the ability to make some controlled vertices to be a bit more spongy. Like the distance holders.
Also fixed a bug with the vertex distance measurer not handling zero distances. It would return NaN which would propagate throughout the whole vertex structure.
Simulation still seems to behave.