Starting working on balance code
I started to work with the centre-of-mass code,  but the existing methods of dealing with balance are in the way and I need tohave a proper think about it all.
There's no body rotation in and the restriction area is now dynamic with the anchor points rather than with the locked position. It now means that the "balancing" status is triggered constantly.