End of day commit
Still trying to get the balance code to work.
I may have to revisit the entire lot.
Tilting regarding balance is sort of working, but it doesn't work in conjunction with sliding to balance. And if continual movement is attempted it results in juddering since the balance sliding is fighting the movement sliding.